+++ xenon-common/src/test/java/com/vmware/xenon/common/TestUtilityService.java
@@ -30,7 +30,7 @@ import com.vmware.xenon.common.Service.ServiceOption;
 import com.vmware.xenon.common.ServiceStats.ServiceStat;
 import com.vmware.xenon.common.ServiceStats.TimeSeriesStats;
 import com.vmware.xenon.common.ServiceStats.TimeSeriesStats.AggregationType;
-import com.vmware.xenon.common.ServiceStats.TimeSeriesStats.DataPoint;
+import com.vmware.xenon.common.ServiceStats.TimeSeriesStats.TimeBin;
 import com.vmware.xenon.common.test.TestContext;
 import com.vmware.xenon.services.common.ExampleService;
 import com.vmware.xenon.services.common.ExampleService.ExampleServiceState;
@@ -309,63 +309,63 @@ public class TestUtilityService extends BasicReusableHostTestCase {
     @Test
     public void testTimeSeriesStats() throws Throwable {
         long startTime = TimeUnit.MILLISECONDS.toMicros(System.currentTimeMillis());
-        int numBuckets = 4;
+        int numBins = 4;
         long interval = 1000;
         double value = 100;
-        // set data to fill up the specified number of buckets
-        TimeSeriesStats timeSeriesStats = new TimeSeriesStats(numBuckets, interval, EnumSet.allOf(AggregationType.class));
-        for (int i = 0; i < numBuckets; i++) {
+        // set data to fill up the specified number of bins
+        TimeSeriesStats timeSeriesStats = new TimeSeriesStats(numBins, interval, EnumSet.allOf(AggregationType.class));
+        for (int i = 0; i < numBins; i++) {
             startTime += TimeUnit.MILLISECONDS.toMicros(interval);
             value += 1;
             timeSeriesStats.add(startTime, value);
         }
-        assertTrue(timeSeriesStats.dataPoints.size() == numBuckets);
+        assertTrue(timeSeriesStats.bins.size() == numBins);
         // insert additional unique datapoints; the earliest entries should be dropped
-        for (int i = 0; i < numBuckets / 2; i++) {
+        for (int i = 0; i < numBins / 2; i++) {
             startTime += TimeUnit.MILLISECONDS.toMicros(interval);
             value += 1;
             timeSeriesStats.add(startTime, value);
         }
-        assertTrue(timeSeriesStats.dataPoints.size() == numBuckets);
-        long timeMicros = startTime - TimeUnit.MILLISECONDS.toMicros(interval * (numBuckets - 1));
+        assertTrue(timeSeriesStats.bins.size() == numBins);
+        long timeMicros = startTime - TimeUnit.MILLISECONDS.toMicros(interval * (numBins - 1));
         long timeMillis = TimeUnit.MICROSECONDS.toMillis(timeMicros);
         timeMillis -= (timeMillis % interval);
-        assertTrue(timeSeriesStats.dataPoints.firstKey() == timeMillis);
+        assertTrue(timeSeriesStats.bins.firstKey() == timeMillis);
 
-        // insert additional datapoints for an existing bucket. The count should increase,
+        // insert additional datapoints for an existing bin. The count should increase,
         // min, max, average computed appropriately
         double origValue = value;
         double accumulatedValue = value;
         double newValue = value;
         double count = 1;
-        for (int i = 0; i < numBuckets / 2; i++) {
+        for (int i = 0; i < numBins / 2; i++) {
             newValue++;
             count++;
             timeSeriesStats.add(startTime, newValue);
             accumulatedValue += newValue;
         }
-        DataPoint lastDatapoint = timeSeriesStats.dataPoints.get(timeSeriesStats.dataPoints.lastKey());
-        assertTrue(lastDatapoint.avg.equals(accumulatedValue / count));
-        assertTrue(lastDatapoint.count == count);
-        assertTrue(lastDatapoint.max.equals(newValue));
-        assertTrue(lastDatapoint.min.equals(origValue));
+        TimeBin lastBin = timeSeriesStats.bins.get(timeSeriesStats.bins.lastKey());
+        assertTrue(lastBin.avg.equals(accumulatedValue / count));
+        assertTrue(lastBin.count == count);
+        assertTrue(lastBin.max.equals(newValue));
+        assertTrue(lastBin.min.equals(origValue));
 
         // test with a subset of the aggregation types specified
-        timeSeriesStats = new TimeSeriesStats(numBuckets, interval, EnumSet.of(AggregationType.AVG));
+        timeSeriesStats = new TimeSeriesStats(numBins, interval, EnumSet.of(AggregationType.AVG));
         timeSeriesStats.add(startTime, value);
-        lastDatapoint = timeSeriesStats.dataPoints.get(timeSeriesStats.dataPoints.lastKey());
-        assertTrue(lastDatapoint.avg != null);
-        assertTrue(lastDatapoint.count != 0);
-        assertTrue(lastDatapoint.max == null);
-        assertTrue(lastDatapoint.min == null);
+        lastBin = timeSeriesStats.bins.get(timeSeriesStats.bins.lastKey());
+        assertTrue(lastBin.avg != null);
+        assertTrue(lastBin.count != 0);
+        assertTrue(lastBin.max == null);
+        assertTrue(lastBin.min == null);
 
-        timeSeriesStats = new TimeSeriesStats(numBuckets, interval, EnumSet.of(AggregationType.MIN, AggregationType.MAX));
+        timeSeriesStats = new TimeSeriesStats(numBins, interval, EnumSet.of(AggregationType.MIN, AggregationType.MAX));
         timeSeriesStats.add(startTime, value);
-        lastDatapoint = timeSeriesStats.dataPoints.get(timeSeriesStats.dataPoints.lastKey());
-        assertTrue(lastDatapoint.avg == null);
-        assertTrue(lastDatapoint.count == 0);
-        assertTrue(lastDatapoint.max != null);
-        assertTrue(lastDatapoint.min != null);
+        lastBin = timeSeriesStats.bins.get(timeSeriesStats.bins.lastKey());
+        assertTrue(lastBin.avg == null);
+        assertTrue(lastBin.count == 0);
+        assertTrue(lastBin.max != null);
+        assertTrue(lastBin.min != null);
 
         // Step 2 - POST a stat to the service instance and verify we can fetch the stat just posted
         String name = UUID.randomUUID().toString();
@@ -381,11 +381,11 @@ public class TestUtilityService extends BasicReusableHostTestCase {
         ServiceStats.ServiceStat stat = new ServiceStat();
         stat.name = "key1";
         stat.latestValue = 100;
-        // set bucket size to 1ms
-        stat.timeSeriesStats = new TimeSeriesStats(numBuckets, 1, EnumSet.allOf(AggregationType.class));
+        // set bin size to 1ms
+        stat.timeSeriesStats = new TimeSeriesStats(numBins, 1, EnumSet.allOf(AggregationType.class));
         this.host.sendAndWaitExpectSuccess(Operation.createPost(UriUtils.buildStatsUri(
                 this.host, exampleServiceState.documentSelfLink)).setBody(stat));
-        for (int i = 0; i < numBuckets; i++) {
+        for (int i = 0; i < numBins; i++) {
             Thread.sleep(1);
             this.host.sendAndWaitExpectSuccess(Operation.createPost(UriUtils.buildStatsUri(
                     this.host, exampleServiceState.documentSelfLink)).setBody(stat));
@@ -394,10 +394,10 @@ public class TestUtilityService extends BasicReusableHostTestCase {
                 UriUtils.buildStatsUri(
                         this.host, exampleServiceState.documentSelfLink));
         ServiceStat retStatEntry = allStats.entries.get(stat.name);
-        assertTrue(retStatEntry.accumulatedValue == 100 * (numBuckets + 1));
+        assertTrue(retStatEntry.accumulatedValue == 100 * (numBins + 1));
         assertTrue(retStatEntry.latestValue == 100);
-        assertTrue(retStatEntry.version == numBuckets + 1);
-        assertTrue(retStatEntry.timeSeriesStats.dataPoints.size() == numBuckets);
+        assertTrue(retStatEntry.version == numBins + 1);
+        assertTrue(retStatEntry.timeSeriesStats.bins.size() == numBins);
 
         // Step 3 - POST a stat to the service instance with sourceTimeMicrosUtc and verify we can fetch the stat just posted
         String statName = UUID.randomUUID().toString();
@@ -423,8 +423,8 @@ public class TestUtilityService extends BasicReusableHostTestCase {
         Long sourceTimeMicrosUtc2 = 946800000000000L;
         sourceStat2.sourceTimeMicrosUtc = sourceTimeMicrosUtc2;
         // set bucket size to 1ms
-        sourceStat1.timeSeriesStats = new TimeSeriesStats(numBuckets, 1, EnumSet.allOf(AggregationType.class));
-        sourceStat2.timeSeriesStats = new TimeSeriesStats(numBuckets, 1, EnumSet.allOf(AggregationType.class));
+        sourceStat1.timeSeriesStats = new TimeSeriesStats(numBins, 1, EnumSet.allOf(AggregationType.class));
+        sourceStat2.timeSeriesStats = new TimeSeriesStats(numBins, 1, EnumSet.allOf(AggregationType.class));
         this.host.sendAndWaitExpectSuccess(Operation.createPost(UriUtils.buildStatsUri(
                 this.host, returnExampleState.documentSelfLink)).setBody(sourceStat1));
         this.host.sendAndWaitExpectSuccess(Operation.createPost(UriUtils.buildStatsUri(
@@ -436,16 +436,16 @@ public class TestUtilityService extends BasicReusableHostTestCase {
         assertTrue(retStatEntry.accumulatedValue == 100);
         assertTrue(retStatEntry.latestValue == 100);
         assertTrue(retStatEntry.version == 1);
-        assertTrue(retStatEntry.timeSeriesStats.dataPoints.size() == 1);
-        assertTrue(retStatEntry.timeSeriesStats.dataPoints.firstKey()
+        assertTrue(retStatEntry.timeSeriesStats.bins.size() == 1);
+        assertTrue(retStatEntry.timeSeriesStats.bins.firstKey()
                 .equals(TimeUnit.MICROSECONDS.toMillis(sourceTimeMicrosUtc1)));
 
         retStatEntry = allStats.entries.get(sourceStat2.name);
         assertTrue(retStatEntry.accumulatedValue == 100);
         assertTrue(retStatEntry.latestValue == 100);
         assertTrue(retStatEntry.version == 1);
-        assertTrue(retStatEntry.timeSeriesStats.dataPoints.size() == 1);
-        assertTrue(retStatEntry.timeSeriesStats.dataPoints.firstKey()
+        assertTrue(retStatEntry.timeSeriesStats.bins.size() == 1);
+        assertTrue(retStatEntry.timeSeriesStats.bins.firstKey()
                 .equals(TimeUnit.MICROSECONDS.toMillis(sourceTimeMicrosUtc2)));
     }